LPS-179567 Extract CommercePriceListQualifiersScreenNavigationCategory methods to a CommercePriceListQualifiersScreenNavigationEntry component and extends CommercePriceListQualifiersScreenNavigationCategory in it
