+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/servlet/taglib/CommercePriceListQualifiersScreenNavigationCategory.java
@@ -14,37 +14,13 @@
 
 package com.liferay.commerce.pricing.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.price.list.model.CommercePriceList;
-import com.liferay.commerce.price.list.service.CommercePriceListAccountRelService;
-import com.liferay.commerce.price.list.service.CommercePriceListChannelRelService;
-import com.liferay.commerce.price.list.service.CommercePriceListCommerceAccountGroupRelService;
-import com.liferay.commerce.price.list.service.CommercePriceListOrderTypeRelService;
-import com.liferay.commerce.price.list.service.CommercePriceListService;
 import com.liferay.commerce.pricing.web.internal.constants.CommercePriceListScreenNavigationConstants;
-import com.liferay.commerce.pricing.web.internal.display.context.CommercePriceListQualifiersDisplayContext;
-import com.liferay.commerce.product.service.CommerceCatalogService;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -53,24 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommercePriceListQualifiersScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommercePriceList> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CommercePriceListScreenNavigationConstants.
-			CATEGORY_KEY_QUALIFIERS;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CommercePriceListScreenNavigationConstants.
 			CATEGORY_KEY_QUALIFIERS;
 	}
@@ -80,7 +46,7 @@ public class CommercePriceListQualifiersScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, "eligibility");
+		return language.get(resourceBundle, "eligibility");
 	}
 
 	@Override
@@ -89,104 +55,7 @@ public class CommercePriceListQualifiersScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_PRICE_LIST_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CommercePriceList commercePriceList) {
-		if ((commercePriceList == null) ||
-			commercePriceList.isCatalogBasePriceList()) {
-
-			return false;
-		}
-
-		boolean hasPermission = false;
-
-		try {
-			hasPermission = _commercePriceListModelResourcePermission.contains(
-				PermissionThreadLocal.getPermissionChecker(),
-				commercePriceList.getCommercePriceListId(), ActionKeys.UPDATE);
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return hasPermission;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		try {
-			CommercePriceListQualifiersDisplayContext
-				commercePriceListQualifiersDisplayContext =
-					new CommercePriceListQualifiersDisplayContext(
-						_commerceCatalogService,
-						_commercePriceListAccountRelService,
-						_commercePriceListChannelRelService,
-						_commercePriceListCommerceAccountGroupRelService,
-						_commercePriceListOrderTypeRelService,
-						_commercePriceListModelResourcePermission,
-						_commercePriceListService, _portal, httpServletRequest);
-
-			httpServletRequest.setAttribute(
-				WebKeys.PORTLET_DISPLAY_CONTEXT,
-				commercePriceListQualifiersDisplayContext);
-		}
-		catch (Exception exception) {
-			_log.error(exception);
-		}
-
-		_jspRenderer.renderJSP(
-			_setServletContext, httpServletRequest, httpServletResponse,
-			"/commerce_price_lists/qualifiers.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CommercePriceListQualifiersScreenNavigationCategory.class);
-
 	@Reference
-	private CommerceCatalogService _commerceCatalogService;
-
-	@Reference
-	private CommercePriceListAccountRelService
-		_commercePriceListAccountRelService;
-
-	@Reference
-	private CommercePriceListChannelRelService
-		_commercePriceListChannelRelService;
-
-	@Reference
-	private CommercePriceListCommerceAccountGroupRelService
-		_commercePriceListCommerceAccountGroupRelService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.price.list.model.CommercePriceList)"
-	)
-	private ModelResourcePermission<CommercePriceList>
-		_commercePriceListModelResourcePermission;
-
-	@Reference
-	private CommercePriceListOrderTypeRelService
-		_commercePriceListOrderTypeRelService;
-
-	@Reference
-	private CommercePriceListService _commercePriceListService;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private Portal _portal;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.pricing.web)"
-	)
-	private ServletContext _setServletContext;
+	protected Language language;
 
 }
